Fix Folder Name, Create "Attributes" Assembly
Split the marker attributes out to their own assembly, as described in
"4. Pack the dll into the generator package" section of the "Solving the
source generator 'marker attribute' problem - Part 2" by Andrew Lock
(https://andrewlock.net/creating-a-source-generator-part-8-solving-the-source-generator-marker-attribute-problem-part2/).

Remove the "IAydskoMapper<>" interface, as it wasn't used and was
causing referencing issues.